# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API do Sistema Financeiro
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/resumo:
    get:
      operationId: getDashboardResumo
      tags: [dashboard]
      summary: Resumo financeiro do dashboard
      parameters:
        - name: mes
          in: query
          schema:
            type: integer
        - name: ano
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Resumo do dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResumo"
  /dashboard/grafico-mensal:
    get:
      operationId: getGraficoMensal
      tags: [dashboard]
      summary: Dados para gráfico de receitas e despesas mensais
      parameters:
        - name: ano
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dados mensais para gráfico
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GraficoMensalItem"
  /dashboard/por-categoria:
    get:
      operationId: getGraficoPorCategoria
      tags: [dashboard]
      summary: Totais agrupados por categoria
      parameters:
        - name: tipo
          in: query
          required: true
          schema:
            type: string
            enum: [receita, despesa]
        - name: mes
          in: query
          schema:
            type: integer
        - name: ano
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Totais por categoria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GraficoCategoriaItem"
  /dashboard/ultimas-transacoes:
    get:
      operationId: getUltimasTransacoes
      tags: [dashboard]
      summary: Últimas transações para o dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Últimas transações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transacao"
...
  /relatorios/exportar:
    get:
      operationId: exportarRelatorio
      tags: [relatorios]
      summary: Exportar relatório em JSON estruturado para geração de Excel/PDF no cliente
      parameters:
        - name: tipo
          in: query
          required: true
          schema:
            type: string
            enum: [transacoes, contasAPagar, contasAReceber, fluxoCaixa]
        - name: dataInicio
          in: query
          schema:
            type: ["string", "null"]
            format: date
        - name: dataFim
          in: query
          schema:
            type: ["string", "null"]
            format: date
        - name: formato
          in: query
          schema:
            type: string
            enum: [excel, pdf]
            default: excel
      responses:
        "200":
          description: Dados estruturados para exportação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelatorioExportData"
components:
  schemas:
...
    Transacao:
      type: object
      required: [id, descricao, valor, tipo, data, status, createdAt]
      properties:
        id:
          type: integer
        descricao:
          type: string
        valor:
          type: number
        tipo:
          type: string
          enum: [receita, despesa]
        data:
          type: string
          format: date
        status:
          type: string
          enum: [pendente, confirmado, cancelado]
        observacoes:
          type: ["string", "null"]
        categoriaId:
          type: ["integer", "null"]
        categoriaNome:
          type: ["string", "null"]
        categoriaIcone:
          type: ["string", "null"]
        categoriaCor:
          type: ["string", "null"]
        contaId:
          type: ["integer", "null"]
        contaNome:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    DashboardResumo:
      type: object
      required: [receitasMes, despesasMes, saldoMes, totalContasAPagarPendente, totalContasAReceberPendente, saldoTotalContas, variacao]
      properties:
        receitasMes:
          type: number
        despesasMes:
          type: number
        saldoMes:
          type: number
        totalContasAPagarPendente:
          type: number
        totalContasAReceberPendente:
          type: number
        saldoTotalContas:
          type: number
        variacao:
          $ref: "#/components/schemas/VariacaoResumo"
    VariacaoResumo:
      type: object
      required: [receitas, despesas, saldo]
      properties:
        receitas:
          type: number
        despesas:
          type: number
        saldo:
          type: number
    GraficoMensalItem:
      type: object
      required: [mes, receitas, despesas, saldo]
      properties:
        mes:
          type: string
        receitas:
          type: number
        despesas:
          type: number
        saldo:
          type: number
    GraficoCategoriaItem:
      type: object
      required: [categoriaId, categoriaNome, cor, icone, total, percentual]
      properties:
        categoriaId:
          type: integer
        categoriaNome:
          type: string
        cor:
          type: string
        icone:
          type: string
        total:
          type: number
        percentual:
          type: number
...
    RelatorioExportData:
      type: object
      required: [tipo, titulo, geradoEm, colunas, linhas]
      properties:
        tipo:
          type: string
        titulo:
          type: string
        geradoEm:
          type: string
          format: date-time
        periodo:
          type: ["string", "null"]
        colunas:
          type: array
          items:
            type: string
        linhas:
          type: array
          items:
            type: object
            additionalProperties: true
        totais:
          type: object
          additionalProperties: true