# lib/api-zod/src/generated/types/dashboardResumo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API do Sistema Financeiro
 * OpenAPI spec version: 0.1.0
 */
import type { VariacaoResumo } from './variacaoResumo';

export interface DashboardResumo {
  receitasMes: number;
  despesasMes: number;
  saldoMes: number;
  totalContasAPagarPendente: number;
  totalContasAReceberPendente: number;
  saldoTotalContas: number;
  variacao: VariacaoResumo;
}