# lib/api-zod/src/generated/types/graficoCategoriaItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API do Sistema Financeiro
 * OpenAPI spec version: 0.1.0
 */

export interface GraficoCategoriaItem {
  categoriaId: number;
  categoriaNome: string;
  cor: string;
  icone: string;
  total: number;
  percentual: number;
}