# lib/api-zod/src/generated/types/variacaoResumo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API do Sistema Financeiro
 * OpenAPI spec version: 0.1.0
 */

export interface VariacaoResumo {
  receitas: number;
  despesas: number;
  saldo: number;
}